sop_deeplang/requirements.txt
# SOP Generation System V6 - Requirements
# 基于LangGraph的多模型SOP自动生成系统

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# LLM Clients
openai>=1.12.0
google-generativeai>=0.5.0

# Data Processing
pydantic>=2.5.0
python-dotenv>=1.0.0

# Optional: For visualization
graphviz>=0.20.0
